refactor(settings_ui): extract theme-prefs persistence to _theme.py
